requirements.txt
pytest==8.3.3
selenium==4.23.1
allure-pytest==2.13.5

allure-python-commons~=2.13.5